build: build socket connection backend
